+++ engine/source/engine.cpp
@@ -3,18 +3,6 @@
 #define STB_IMAGE_IMPLEMENTATION
 #include <stb_image.h>
 
-#pragma clang diagnostic push
-#pragma clang diagnostic ignored "-Wnullability-completeness"
-#pragma clang diagnostic ignored "-Wunused-variable"
-#define VMA_IMPLEMENTATION
-#define VMA_LEAK_LOG_FORMAT(format, ...) do { \
-        spdlog::info(fmt::sprintf(format, __VA_ARGS__)); \
-    } while(false)
-
-#include "vk_mem_alloc.h"
-
-#pragma clang diagnostic pop
-
 #include "vulkan_validation.hpp"
 #include "vulkan_helper.hpp"
 #include "imgui_impl_vulkan.h"
@@ -152,7 +140,7 @@ void Engine::Run()
         constexpr glm::vec3 RIGHT = { 1.0f, 0.0f, 0.0f};
         constexpr glm::vec3 FORWARD = { 0.0f, 0.0f, 1.0f };
         //constexpr glm::vec3 UP = { 0.0f, -1.0f, 0.0f };
-        
+
         _scene.camera.euler_rotation.x -= mouse_delta.y * MOUSE_SENSITIVITY;
         _scene.camera.euler_rotation.y -= mouse_delta.x * MOUSE_SENSITIVITY;
