+++ engine/source/vma_impl.cpp
@@ -1,0 +1,11 @@
+#pragma clang diagnostic push
+#pragma clang diagnostic ignored "-Wnullability-completeness"
+#pragma clang diagnostic ignored "-Wunused-variable"
+#define VMA_IMPLEMENTATION
+#define VMA_LEAK_LOG_FORMAT(format, ...) do { \
+spdlog::info(fmt::sprintf(format, __VA_ARGS__)); \
+    } while(false)
+
+#include "vk_mem_alloc.h"
+
+#pragma clang diagnostic pop
